utils for parsing config json
