+++ packages/amplify_flutter/lib/utils/parse_json_config.dart
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+import 'package:amplify_flutter/config/amplify_config.dart';
+import 'package:amplify_flutter/utils/create_config_object.dart';
+import 'package:flutter/foundation.dart';
+
+Future<AmplifyConfig?> parseConfigJson(String configuration) async {
+  try {
+    return createConfigObject(configuration);
+  } catch (e) {
+    if (kDebugMode) {
+      debugPrint(
+          'There was an unexpected problem parsing the amplifyconfiguration.dart file.');
+    }
+  }
+}
